refactor: move cel helper fn to better suited pkgs
